removed duplicate logic
reworked api routes
